converting example to use non xml configs
