+++ examples/rollbar-spring-webmvc/.gitignore
@@ -1,4 +1,3 @@
 .idea
-*.iml
 target
-*.DS_Store
+*.iml
