+++ examples/rollbar-spring-webmvc/src/main/java/com/example/springwebmvc/initializer/AppInitializer.java
@@ -1,0 +1,27 @@
+package com.example.springwebmvc.initializer;
+
+import com.example.springwebmvc.config.RollbarConfig;
+import org.springframework.web.WebApplicationInitializer;
+import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
+import org.springframework.web.servlet.DispatcherServlet;
+
+import javax.servlet.ServletContext;
+import javax.servlet.ServletException;
+import javax.servlet.ServletRegistration;
+
+public class AppInitializer implements WebApplicationInitializer {
+
+    @Override
+    public void onStartup(ServletContext servletContext) throws ServletException {
+        AnnotationConfigWebApplicationContext ctx = new AnnotationConfigWebApplicationContext();
+        ctx.register(RollbarConfig.class);
+        ctx.setServletContext(servletContext);
+        ctx.refresh();
+
+        ServletRegistration.Dynamic servlet = servletContext.addServlet("dispatcher", new DispatcherServlet(ctx));
+        servlet.setLoadOnStartup(1);
+        servlet.addMapping("/");
+
+    }
+
+}
