--- examples/rollbar-spring-webmvc/src/main/resources/application.properties
@@ -1,3 +1,0 @@
-rollbar.access_token = 1b7dec62e84341ff8361b91d2c94e5b4
-rollbar.environment = development
-rollbar.framework = spring-webmvc
