--- examples/rollbar-spring-webmvc/src/main/resources/logback.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE logback>
-<configuration>
-
-    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>%date{HH:mm:ss.SSS} [%thread] %-5level %logger{15}#%line %msg\n</pattern>
-        </encoder>
-    </appender>
-
-    <logger name="com.example.springwebmvc">
-        <level value="debug"/>
-    </logger>
-
-    <logger name="org.springframework">
-        <level value="info"/>
-    </logger>
-
-    <root>
-        <level value="error"/>
-        <appender-ref ref="STDOUT"/>
-    </root>
-
-</configuration>
