--- examples/rollbar-spring-webmvc/src/main/webapp/WEB-INF/appconfig-root.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:context="http://www.springframework.org/schema/context"
-       xmlns="http://www.springframework.org/schema/beans"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
-
-		http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd">
-
-    <import resource="appconfig-mvc.xml"/>
-
-    <!-- Scans within the base package of the application for @Component classes to configure as beans -->
-    <context:component-scan base-package="com.example.springwebmvc"/>
-
-    <context:property-placeholder location="classpath:application.properties"/>
-
-</beans>
